magnetometer integrated into adcs
