+++ src/tasks/adcs/adcs_main.c
@@ -37,6 +37,13 @@ QueueHandle_t init_adcs(void) {
         warning("photodiode: Hardware initialization failed\n");
     }
 
+    // Initialize magnetometer hardware
+    info("Initializing magnetometer\n");
+    result = init_rm3100();
+    if (result != SUCCESS) {
+        warning("rm3100: Hardware initialization failed\n");
+    }
+
     // Initialize RTC timer hardware
     info("Initializing RTC timer\n");
     result = init_rtc_hardware();
