+++ src/tasks/adcs/photodiode_driver.c
@@ -7,6 +7,7 @@
  * Modified: November 9, 2025
  * Authors: Avinash Patel, Yi Lyo, Alexander Thaep
  */
+
 #include <inttypes.h>
 #include "photodiode_driver.h"
 #include "adcs_task.h"
@@ -25,6 +26,42 @@ status_t init_photodiode_hardware(void) {
     // TODO: Initialize any required peripherals
 
     info("photodiode_driver: Hardware initialization complete\n");
+    return SUCCESS;
+}
+
+/**
+ * \fn photodiode_read
+ *
+ * \brief Reads photodiode values and calculates sun vector
+ *
+ * \param data pointer to photodiode_data_t structure to fill
+ *
+ * \returns status_t SUCCESS if reading was successful
+ */
+status_t photodiode_read(photodiode_data_t *const data) {
+    if (!data) {
+        return ERROR_SANITY_CHECK_FAILED;
+    }
+
+    debug("photodiode: Reading photodiode values\n");
+
+    // Read raw ADC values
+    uint16_t raw_values[PHOTODIODE_COUNT];
+    status_t result = read_photodiodes(raw_values);
+
+    if (result != SUCCESS) {
+        warning("photodiode: ADC read failed\n");
+        return result;
+    }
+
+    // Copy raw values to data structure
+    for (int i = 0; i < PHOTODIODE_COUNT; i++) {
+        data->raw_values[i] = raw_values[i];
+    }
+
+    data->timestamp = xTaskGetTickCount();
+    data->valid = true;
+
     return SUCCESS;
 }
