+++ src/tasks/task_list.h
@@ -6,7 +6,6 @@
 #include "display_task.h"
 #include "globals.h"
 #include "heartbeat_task.h"
-#include "magnetometer_task.h"
 #include "adcs_task.h"
 #include "shell_task.h"
 #include "task_manager_task.h"
@@ -16,7 +15,6 @@
 extern pvdx_task_t *const p_watchdog_task;
 extern pvdx_task_t *const p_command_dispatcher_task;
 extern pvdx_task_t *const p_task_manager_task;
-extern pvdx_task_t *const p_magnetometer_task;
 extern pvdx_task_t *const p_adcs_task;
 extern pvdx_task_t *const p_shell_task;
 extern pvdx_task_t *const p_display_task;
